refactor : common TokenVariant between direct and bytecode interpreters

+++ interpreter/src/Interpreter/ComputingOperations/InterpretNumeric.cpp
@@ -9,17 +9,17 @@ std::string ska::nodeval_to_string<ska::StringShared>(const StringShared& obj) {
 ska::NodeValue ska::InterpretMathematicPlus(TypedNodeValue firstValue, TypedNodeValue secondValue, const Type& destinationType) {
 	switch (destinationType.type()) {
 	case ExpressionType::STRING: {
-		return ComputeTwoTypeOperation<int, StringShared>(std::move(firstValue), std::move(secondValue),
+		return ComputeTwoTypeOperation<long, StringShared>(std::move(firstValue), std::move(secondValue),
 		[](const auto& t1, const auto& t2) {
 			return std::make_shared<std::string>((nodeval_to_string(t1) + nodeval_to_string(t2)));
 		});
 	}
 
 	case ExpressionType::INT:
-		return firstValue.value.nodeval<int>() + secondValue.value.nodeval<int>();
+		return firstValue.value.nodeval<long>() + secondValue.value.nodeval<long>();
 
 	case ExpressionType::FLOAT: {
-		return ComputeTwoTypeOperation<int, double>(std::move(firstValue), std::move(secondValue),
+		return ComputeTwoTypeOperation<long, double>(std::move(firstValue), std::move(secondValue),
 		[](const auto & t1, const auto & t2) {
 			return t1 + t2;
 		});
@@ -44,7 +44,7 @@ ska::NodeValue ska::InterpretMathematicPlus(TypedNodeValue firstValue, TypedNode
 			std::get<2>(lambdas)
 			);
 		case ExpressionType::INT:
-			return ComputeTwoTypeOperation<int, NodeValueArray>(std::move(firstValue), std::move(secondValue),
+			return ComputeTwoTypeOperation<long, NodeValueArray>(std::move(firstValue), std::move(secondValue),
 			std::get<0>(lambdas),
 			std::get<1>(lambdas),
 			std::get<2>(lambdas));
@@ -67,20 +67,20 @@ ska::NodeValue ska::InterpretMathematicPlus(TypedNodeValue firstValue, TypedNode
 ska::NodeValue ska::InterpretMathematicMinus(TypedNodeValue firstValue, TypedNodeValue secondValue, const Type& destinationType) {
 	switch (destinationType.type()) {
 	case ExpressionType::INT:
-		return firstValue.value.nodeval<int>() - secondValue.value.nodeval<int>();
+		return firstValue.value.nodeval<long>() - secondValue.value.nodeval<long>();
 
 	case ExpressionType::FLOAT: {
-		return ComputeTwoTypeOperation<int, double>(std::move(firstValue), std::move(secondValue),
+		return ComputeTwoTypeOperation<long, double>(std::move(firstValue), std::move(secondValue),
 		[](const auto& t1, const auto& t2) {
 			return t1 + t2;
 		});
 	}
 
 	case ExpressionType::ARRAY: {
-		auto lambdas = std::make_tuple([](int t1, NodeValueArray& t2) {
+		auto lambdas = std::make_tuple([](long t1, NodeValueArray& t2) {
 			throw std::runtime_error("Unhandled math minus operation");
 			return t2;
-		}, [](NodeValueArray& t1, int t2) {
+		}, [](NodeValueArray& t1, long t2) {
 			if(t2 < 0) {
 				throw std::runtime_error("Bad vector-based index deletion : negative index(es)");
 			}
@@ -97,7 +97,7 @@ ska::NodeValue ska::InterpretMathematicMinus(TypedNodeValue firstValue, TypedNod
 
 		switch (secondValue.type.type()) {
 		case ExpressionType::INT:
-			return ComputeTwoTypeOperation<int, NodeValueArray>(std::move(firstValue), std::move(secondValue),
+			return ComputeTwoTypeOperation<long, NodeValueArray>(std::move(firstValue), std::move(secondValue),
 			std::get<0>(lambdas),
 			std::get<1>(lambdas),
 			std::get<2>(lambdas));
@@ -107,11 +107,11 @@ ska::NodeValue ska::InterpretMathematicMinus(TypedNodeValue firstValue, TypedNod
 			if(t2->size() != 2) {
 				throw std::runtime_error("Bad vector-based index deletion : size != 2");
 			}
-			if((*t2)[0].nodeval<int>() < 0 || (*t2)[1].nodeval<int>() < 0) {
+			if((*t2)[0].nodeval<long>() < 0 || (*t2)[1].nodeval<long>() < 0) {
 				throw std::runtime_error("Bad vector-based index deletion : negative index(es)");
 			}
-			auto start = static_cast<std::size_t>((*t2)[0].nodeval<int>());
-			auto end = static_cast<std::size_t>((*t2)[1].nodeval<int>());
+			auto start = static_cast<std::size_t>((*t2)[0].nodeval<long>());
+			auto end = static_cast<std::size_t>((*t2)[1].nodeval<long>());
 			if(start > end) {
 				std::swap(start, end);
 			}
@@ -139,7 +139,7 @@ ska::NodeValue ska::InterpretMathematicMinus(TypedNodeValue firstValue, TypedNod
 ska::NodeValue ska::InterpretMathematicMultiply(TypedNodeValue firstValue, TypedNodeValue secondValue, const Type& destinationType) {
 	switch (destinationType.type()) {
 	case ExpressionType::STRING: {
-		return ComputeTwoTypeOperation<int, StringShared>(std::move(firstValue), std::move(secondValue),
+		return ComputeTwoTypeOperation<long, StringShared>(std::move(firstValue), std::move(secondValue),
 		[](int t1, const StringShared& t2) {
 			assert(t1 > 0);
 			auto ss = std::stringstream{};
@@ -161,10 +161,10 @@ ska::NodeValue ska::InterpretMathematicMultiply(TypedNodeValue firstValue, Typed
 	}
 
 	case ExpressionType::INT:
-		return firstValue.value.nodeval<int>() * secondValue.value.nodeval<int>();
+		return firstValue.value.nodeval<long>() * secondValue.value.nodeval<long>();
 
 	case ExpressionType::FLOAT: {
-		return ComputeTwoTypeOperation<int, double>(std::move(firstValue), std::move(secondValue),
+		return ComputeTwoTypeOperation<long, double>(std::move(firstValue), std::move(secondValue),
 		[](const auto& t1, const auto& t2) {
 			return t1 * t2;
 		});
@@ -179,13 +179,13 @@ ska::NodeValue ska::InterpretMathematicMultiply(TypedNodeValue firstValue, Typed
 ska::NodeValue ska::InterpretMathematicDivide(TypedNodeValue firstValue, TypedNodeValue secondValue, const Type& destinationType) {
 	switch (destinationType.type()) {
 	case ExpressionType::INT:
-		if (secondValue.value.nodeval<int>() == 0) {
+		if (secondValue.value.nodeval<long>() == 0) {
 			throw std::runtime_error("math error : division by a null value");
 		}
-		return firstValue.value.nodeval<int>() / secondValue.value.nodeval<int>();
+		return firstValue.value.nodeval<long>() / secondValue.value.nodeval<long>();
 
 	case ExpressionType::FLOAT: {
-		return ComputeTwoTypeOperation<int, double>(std::move(firstValue), std::move(secondValue),
+		return ComputeTwoTypeOperation<long, double>(std::move(firstValue), std::move(secondValue),
 		[](const auto& t1, const auto& t2) {
 			if (t2 == 0) {
 				throw std::runtime_error("math error : division by a null value");
