+++ interpreter/src/Interpreter/InterpreterOperatorArrayUse.cpp
@@ -7,7 +7,7 @@
 ska::NodeCell ska::InterpreterOperator<ska::Operator::ARRAY_USE>::interpret(OperateOn node) {
 	auto array = m_interpreter.interpret({ node.parent, node.GetArrayNode() }).asLvalue();
 	auto& arrayCell = *array.object->as<NodeValueArray>();
-	auto arrayIndex = m_interpreter.interpret({ node.parent, node.GetArrayIndexNode() }).asRvalue().object.nodeval<int>();
+	auto arrayIndex = m_interpreter.interpret({ node.parent, node.GetArrayIndexNode() }).asRvalue().object.nodeval<long>();
 	if (arrayCell.size() <= arrayIndex) {
 		throw std::runtime_error("array index out of bounds");
 	}
