+++ src/src/NodeValue/TypeFromNative.cpp
@@ -1,6 +1,6 @@
 #include "TypeFromNative.h"
 
-const ska::Type ska::TypeFromNative<int>::value() { return ska::Type::MakeBuiltIn<ska::ExpressionType::INT>(); }
+const ska::Type ska::TypeFromNative<long>::value() { return ska::Type::MakeBuiltIn<ska::ExpressionType::INT>(); }
 const ska::Type ska::TypeFromNative<float>::value() { return ska::Type::MakeBuiltIn<ska::ExpressionType::FLOAT>(); }
 const ska::Type ska::TypeFromNative<double>::value() { return ska::Type::MakeBuiltIn<ska::ExpressionType::FLOAT>(); }
 const ska::Type ska::TypeFromNative<bool>::value() { return ska::Type::MakeBuiltIn<ska::ExpressionType::BOOLEAN>(); }
