+++ std/src/std/module/function/parameter.cpp
@@ -15,17 +15,17 @@ ska::lang::ParameterModule::ParameterModule(ModuleConfiguration& config, const s
         auto result = m_bridge.createMemory();
         result->emplace("asInt", std::make_unique<ska::BridgeFunction>(
             std::function<ska::NodeValue(std::vector<ska::NodeValue>)>([&](std::vector<ska::NodeValue> params) {
-            const auto index = params[0].nodeval<int>();
-            return m_parameters.size() > index ? static_cast<int>(m_parameters[index].convertNumeric()) : -1;
+            const auto index = params[0].nodeval<long>();
+            return m_parameters.size() > index ? static_cast<long>(m_parameters[index].convertNumeric()) : -1;
         })));
         result->emplace("asString", std::make_unique<ska::BridgeFunction>(
             std::function<ska::NodeValue(std::vector<ska::NodeValue>)>([&](std::vector<ska::NodeValue> params) {
-            const auto index = params[0].nodeval<int>();
+            const auto index = params[0].nodeval<long>();
             return std::make_shared<std::string>(std::move(m_parameters.size() > index ? m_parameters[index].convertString() : ""));
         })));
         result->emplace("size", std::make_unique<ska::BridgeFunction>(
             std::function<ska::NodeValue(std::vector<ska::NodeValue>)>([&](std::vector<ska::NodeValue> params) {
-            return static_cast<int>(m_parameters.size());
+            return static_cast<long>(m_parameters.size());
         })));
         return result;
     }));
